For #3384, Move common utils and DataSourcePropertiesSetter to new module
