+++ sharding-spring/sharding-jdbc-spring/sharding-jdbc-spring-boot-starter/pom.xml
@@ -28,6 +28,11 @@
     
     <dependencies>
         <dependency>
+            <groupId>org.apache.shardingsphere</groupId>
+            <artifactId>sharding-spring-boot-util</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter</artifactId>
         </dependency>
