+++ sharding-spring/sharding-spring-boot-util/src/main/java/org/apache/shardingsphere/spring/boot/util/DataSourceUtil.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.shardingjdbc.orchestration.spring.boot.util;
+package org.apache.shardingsphere.spring.boot.util;
 
 import com.google.common.base.CaseFormat;
 import com.google.common.collect.Sets;
