+++ sharding-spring/sharding-spring-boot-util/src/main/resources/META-INF/services/org.apache.shardingsphere.spring.boot.datasource.DataSourcePropertiesSetter
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.shardingjdbc.spring.boot.datasource.HikariDataSourcePropertiesSetter
+org.apache.shardingsphere.spring.boot.datasource.HikariDataSourcePropertiesSetter
