+++ sharding-spring/sharding-spring-boot-util/src/test/java/org/apache/shardingsphere/spring/boot/util/DataSourceUtilTest.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.shardingjdbc.orchestration.spring.boot.util;
+package org.apache.shardingsphere.spring.boot.util;
 
 import com.zaxxer.hikari.HikariDataSource;
 import org.apache.commons.dbcp2.BasicDataSource;
